fix: Resolve MSVC compilation warnings for Windows builds
- Fix variable shadowing warnings (C4456):
  - main.cpp: Remove duplicate table_mgr declaration at line 882
  - main.cpp: Rename shadowed meta variable to random_meta at line 1500

- Fix size_t to int conversion warnings (C4267):
  - main.cpp: Add static_cast for all_points.size()
  - terminal.cpp: Add static_cast for buffer size conversions
  - axis_renderer.cpp: Add static_cast for label length and loop indices (4 locations)
  - header_renderer.cpp: Add static_cast for string length and loop indices (5 locations)
  - table_creation_dialog.cpp: Add static_cast for string operations (9 locations)

- Fix int64 to unsigned int conversion warning (C4244):
  - random_initializer.cpp: Add static_cast for RNG seed

All warnings were treated as errors by MSVC /WX flag.
These fixes ensure the code compiles cleanly on Windows.

+++ src/axis_renderer.cpp
@@ -228,11 +228,11 @@ void AxisRenderer::render_x_axis(Terminal& terminal, const Viewport& viewport,
 
                 // Draw label (centered on tick)
                 // Place label characters around the tick position
-                int label_len = tick.label.length();
+                int label_len = static_cast<int>(tick.label.length());
                 int label_start = tick_col - label_len / 2;
 
                 for (size_t i = 0; i < tick.label.length(); ++i) {
-                    int label_col = label_start + i;
+                    int label_col = label_start + static_cast<int>(i);
                     if (label_col >= start_col && label_col < start_col + width &&
                         label_col != tick_col) {  // Don't overwrite tick mark
                         terminal.write_char(axis_row, label_col, tick.label[i]);
@@ -279,12 +279,12 @@ void AxisRenderer::render_y_axis(Terminal& terminal, const Viewport& viewport,
 
                 // Draw label (to the left of the axis)
                 // Place label characters to the left of the tick
-                int label_len = tick.label.length();
+                int label_len = static_cast<int>(tick.label.length());
                 int label_start = axis_col - label_len - 1;
 
                 if (label_start >= 0) {
                     for (size_t i = 0; i < tick.label.length(); ++i) {
-                        int label_col = label_start + i;
+                        int label_col = label_start + static_cast<int>(i);
                         if (label_col >= 0) {
                             terminal.write_char(tick_row, label_col, tick.label[i]);
                         }
